Refactor into modular rfp_app package
